@@ -5,8 +5,8 @@ def safeExtGet(prop, fallback) {
 }
 
 android {
-    compileSdkVersion safeExtGet('compileSdkVersion', 27)
-    buildToolsVersion safeExtGet('buildToolsVersion', '27.0.3')
+    compileSdkVersion safeExtGet('compileSdkVersion', 28)
+    buildToolsVersion safeExtGet('buildToolsVersion', '28.0.3')
  
     compileOptions {
         targetCompatibility JavaVersion.VERSION_1_8
@@ -15,7 +15,7 @@ android {
 
     defaultConfig {
         minSdkVersion safeExtGet('minSdkVersion', 16)
-        targetSdkVersion safeExtGet('targetSdkVersion', 27)
+        targetSdkVersion safeExtGet('targetSdkVersion', 28)
         versionCode 1
         versionName "1.0"
     }
@@ -28,9 +28,12 @@ dependencies {
     }
 
     // All support libs must use the same version
-    implementation "com.android.support:support-annotations:${safeExtGet('supportLibVersion', '+')}"
-    implementation "com.android.support:support-compat:${safeExtGet('supportLibVersion', '+')}"
-    implementation "com.android.support:support-media-compat:${safeExtGet('supportLibVersion', '+')}"
+    // implementation "com.android.support:support-annotations:${safeExtGet('supportLibVersion', '+')}"
+    implementation "androidx.annotation:annotation:1.0.0"
+    // implementation "com.android.support:support-compat:${safeExtGet('supportLibVersion', '+')}"
+    implementation "androidx.core:core:1.0.0"
+    // implementation "com.android.support:support-media-compat:${safeExtGet('supportLibVersion', '+')}"
+    implementation "androidx.media:media:1.0.0"
 
     implementation('com.google.android.exoplayer:extension-okhttp:2.9.3') {
         exclude group: 'com.squareup.okhttp3', module: 'okhttp'
