@@ -83,13 +83,16 @@ def enableSeparateBuildPerCPUArchitecture = false
 def enableProguardInReleaseBuilds = false
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion '25.0.2'
-
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
     defaultConfig {
         applicationId "com.videoplayer"
-        minSdkVersion 16
-        targetSdkVersion 25
+        minSdkVersion 19
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
         ndk {
@@ -108,6 +111,7 @@ android {
         release {
             signingConfig signingConfigs.debug
             minifyEnabled enableProguardInReleaseBuilds
+            matchingFallbacks = ['release', 'debug']
             proguardFiles getDefaultProguardFile("proguard-android.txt"), "proguard-rules.pro"
         }
     }
@@ -127,9 +131,11 @@ android {
 }
 
 dependencies {
-    compile "com.android.support:appcompat-v7:25.2.0"
-    compile "com.facebook.react:react-native:+"  // From node_modules
-    compile project(':react-native-video')
+    implementation project(':react-native-video')
+    // compile "com.android.support:appcompat-v7:25.2.0"
+    implementation "androidx.appcompat:appcompat:1.0.0"
+    implementation "com.facebook.react:react-native:+"  // From node_modules
+    // implementation project(':react-native-video')
 //    compile project(':react-native-video-exoplayer') // uncomment to use exoplayer
 }
 
